requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
boto3==1.18.40
    # via -r requirements.in
botocore==1.21.40
    # via
    #   boto3
    #   s3transfer
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
python-dateutil==2.8.2
    # via botocore
python-dotenv==0.19.0
    # via -r requirements.in
s3transfer==0.5.0
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.6
    # via botocore
